+++ src/layout/LayoutBuilder.tsx
@@ -186,40 +186,3 @@ export const LayoutBuilder: FC<Props> = memo(
     );
   }
 );
-
-/**
- * 
- * 
-        @media only screen and (max-width: 1195px) {
-          #page {
-            grid-template-areas:
-              "header header"
-              "main main"
-              "main main"
-              "main main";
-          }
-
-          nav {
-            position: fixed;
-            left: 0;
-            bottom: 0;
-            top: ${Dimensions.HEADER_HEIGHT}px;
-            height: calc(100vh - ${Dimensions.HEADER_HEIGHT}px);
-            transform: translateX(-100%);
-            transition: transform 0.3s;
-            z-index: 1000;
-          }
-
-          main {
-            height: auto;
-          }
-
-          .open {
-            transform: translateX(0px);
-          }
-
-          .closed {
-            transform: translateX(-100%);
-          }
-        }
- */
