+++ src/pages/blog/jumpthrow-bind.tsx
@@ -3,6 +3,8 @@ import { LayoutBuilder } from "../../layout/LayoutBuilder";
 import { JumpThrowBindBlogMain } from "../../blog/blogposts/JumpThrowBind";
 import { HeaderDefault } from "../../defaultheader/Header";
 import { Navigation } from "../../navigation/Navigation";
+import { BlogDefaultSidebar } from "../../blog/BlogDefaultSidebar";
+import { blogJumpthrowBind } from "../../blog/ArticleData/blogPosts";
 
 const NadeAlignCrosshairBlogPost: FC = () => {
   return (
@@ -10,6 +12,7 @@ const NadeAlignCrosshairBlogPost: FC = () => {
       nav={<Navigation />}
       header={<HeaderDefault />}
       main={<JumpThrowBindBlogMain></JumpThrowBindBlogMain>}
+      sidebar={<BlogDefaultSidebar data={blogJumpthrowBind} />}
     />
   );
 };
