+++ src/pages/blog/practice-config.tsx
@@ -2,6 +2,8 @@ import { LayoutBuilder } from "../../layout/LayoutBuilder";
 import { Navigation } from "../../navigation/Navigation";
 import { HeaderDefault } from "../../defaultheader/Header";
 import { PractiseConfigBlogPost } from "../../blog/blogposts/PracticeConfig";
+import { BlogDefaultSidebar } from "../../blog/BlogDefaultSidebar";
+import { blogPractiseConfig } from "../../blog/ArticleData/blogPosts";
 
 const PractiseConfigBlogPostPage = () => {
   return (
@@ -9,6 +11,7 @@ const PractiseConfigBlogPostPage = () => {
       nav={<Navigation />}
       header={<HeaderDefault />}
       main={<PractiseConfigBlogPost />}
+      sidebar={<BlogDefaultSidebar data={blogPractiseConfig} />}
     />
   );
 };
