+++ src/pages/blog/smoke-align-crosshair.tsx
@@ -1,8 +1,10 @@
 import { FC } from "react";
+import { BlogDefaultSidebar } from "../../blog/BlogDefaultSidebar";
 import { NadeAlignCrosshairPost } from "../../blog/blogposts/NadeAlignCrosshairPost";
 import { HeaderDefault } from "../../defaultheader/Header";
 import { LayoutBuilder } from "../../layout/LayoutBuilder";
 import { Navigation } from "../../navigation/Navigation";
+import { blogNadeAlignCrosshair } from "../../blog/ArticleData/blogPosts";
 
 const NadeAlignCrosshairBlogPost: FC = () => {
   return (
@@ -10,6 +12,7 @@ const NadeAlignCrosshairBlogPost: FC = () => {
       nav={<Navigation />}
       header={<HeaderDefault />}
       main={<NadeAlignCrosshairPost />}
+      sidebar={<BlogDefaultSidebar data={blogNadeAlignCrosshair} />}
     />
   );
 };
